requirements.txt
fastapi
uvicorn
anthropic
slowapi
gspread
google-auth
